services/anomaly-detection-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
kafka-python==2.0.2
python-dotenv==1.0.1
